games_ratings_consumer/requirements.txt
pymongo>=4.2.0,<4.3.0
python-dotenv==0.20.0
pika>=1.3.0,<1.4.0
